// src/app/models/scraped-recipe.ts
export class ScrapedRecipe {
  name: string;
  alternateName: string;
  URL: string;
  totalTime: string;
  image: Array<string>;
  description: string;
  recipeYield: string;
  recipeCategory: string;
  recipeCuisine: string;
  recipeIngredient: Array<string>;
  recipeInstructions: Array<string>;
}